+++ packages/contracts/scripts/tasks/nodeTasks.ts
@@ -1,7 +1,5 @@
 import { task, types } from 'hardhat/config'
 import { minStateDuration } from 'scripts/node/deploy'
-import { BigNumber, utils } from 'ethers'
-import { EthereumProvider } from 'hardhat/types'
 
 task('increase-time', 'Increases block time')
   .addParam('value', 'Time in seconds to increase', minStateDuration, types.int, true)
@@ -21,34 +19,3 @@ task('accounts', 'Prints available accounts')
     const signers = await hre.ethers.getSigners()
     signers.forEach((signer, index) => console.log(`Account #${index} ${signer.address}`))
   })
-
-interface Block {
-  hash: string,
-}
-
-type GetBlockNumberResponse = Block | undefined | null
-
-task('generate-random-numbers', 'Generates random numbers')
-  .addParam('blocks', 'Array of block numbers', undefined, types.json)
-  .addParam('secret', 'Secret number', undefined, types.int)
-  .setAction(async ({ blocks, secret }: { blocks: number[], secret: number }, hre) => {
-    const provider = hre.network.provider
-    const hexSecret = BigNumber.from(secret).toHexString().substring(2)
-    const randomNumbers: BigNumber[] = []
-
-    for (const block of blocks) {
-      const blockHash = await getBlockHash(provider, block)
-      const randomNumberHash = utils.keccak256(blockHash.concat(hexSecret))
-      randomNumbers.push(BigNumber.from(randomNumberHash))
-    }
-
-    randomNumbers.forEach((randomNumber, index) => {
-      console.log(`Random number for block #${blocks[index]}: ${randomNumber}`)
-    })
-  })
-
-async function getBlockHash(provider: EthereumProvider, blockNumber: number) {
-  const hexBlockNumber = BigNumber.from(blockNumber).toHexString()
-  const block: GetBlockNumberResponse = await provider.send('eth_getBlockByNumber', [hexBlockNumber, true])
-  return block.hash
-}
